#76 refactor: 패키지 분리
